RMB-934: Vert.x 4.3.3 fixing disabled SSL in 4.3.0/4.3.1
Vert.x 4.3.0 and 4.3.1 have this regression: They disable SSL in WebClient.

Vert.x >= 4.3.2 has a fix.

https://github.com/vert-x3/vertx-web/issues/2196 "SSL not set for secure Request"

https://github.com/vert-x3/vertx-web/pull/2226 "WebClient ignores the SSL flag when a request is created from request options with the SSL flag true"

https://github.com/vert-x3/vertx-web/issues/2225 "HttpRequestImpl's SSL Isn't Set By RequestOptions"
